requirements.txt
openai==0.28.0
transformer==4.35.0
dashscope==1.13.3
scikit-learn=1.3.2
torch==2.3.0
